Move the search functionality to its own file
